initial commit
Set the date displayed in the calendar. Wrote the calendar to the element with the id "calendar. Added the function to generate calendar table. Made the Function to write the calendar caption. Added month array contains the list of month names. Added the current month and year. Added the caption with the month and year.