Final commit / ready for review
Added the script for the event javascript in the html file. wrote the cells for each day of the month. wrote the blank cells preceding the starting day. Added the function to write the table rows for each day of the month. Wrote variables to determine the starting day of the month. return the number of days for the current month. Ready for review.